+++ mne/io/ctf/info.py
@@ -6,12 +6,14 @@
 
 from time import strptime
 from calendar import timegm
+import os.path as op
 
 import numpy as np
 
-from ...utils import logger, warn
+from ...utils import logger, warn, _clean_names
 from ...transforms import (apply_trans, _coord_frame_name, invert_transform,
                            combine_transforms)
+from ...annotations import Annotations
 
 from ..meas_info import _empty_info
 from ..write import get_new_file_id
@@ -448,3 +450,35 @@ def _compose_meas_info(res4, coils, trans, eeg):
     logger.info('    Measurement info composed.')
     info._update_redundant()
     return info
+
+
+def _read_bad_chans(directory, info):
+    """Read Bad channel list and match to internal names."""
+    fname = op.join(directory, 'BadChannels')
+    if not op.exists(fname):
+        return []
+    mapping = dict(zip(_clean_names(info['ch_names']), info['ch_names']))
+    with open(fname, 'r') as fid:
+        bad_chans = [mapping[f.strip()] for f in fid.readlines()]
+    return bad_chans
+
+
+def _annotate_bad_segments(directory, start_time):
+    fname = op.join(directory, 'bad.segments')
+    if not op.exists(fname):
+        return None
+    # read in bad segment file
+    onsets = []
+    durations = []
+    desc = []
+    with open(fname, 'r') as fid:
+        for f in fid.readlines():
+            tmp = f.strip().split()
+            desc.append('bad_%s' % tmp[0])
+            onsets.append(np.float(tmp[1]) - start_time)
+            durations.append(np.float(tmp[2]) - np.float(tmp[1]))
+    # return None if there are no bad segments
+    if len(onsets) == 0:
+        return None
+
+    return Annotations(onsets, durations, desc)