+++ doc/whats_new.rst
@@ -62,6 +62,8 @@ Bug
 
 - Fix error when interpolating MEG channels with compensation using reference channels (like for CTF data) by `Alex Gramfort`_
 
+- Fix bug in :func:`mne.io.ctf.read_raw_ctf` to read bad channels and segments from CTF ds files by `Luke Bloy`_
+
 - Fix problem with :meth:`mne.io.Raw.add_channels` where ``raw.info['bads']`` was replicated by `Eric Larson`_
 
 - Rename ``raw.info['buffer_size_sec']`` to ``raw.buffer_size_sec`` as it is a writing parameter rather than a measurement parameter by `Eric Larson`_